+++ app/(payment)/cancel.jsx
@@ -1,0 +1,12 @@
+const SuccesFailed = () => {
+  return (
+    <>
+      <WebView
+        source={{ uri: "https://www.facebook.com/" }}
+        style={{ flex: 1 }}
+      />
+    </>
+  );
+};
+
+export default SuccesFailed;
