+++ app/subSite/create.tsx
@@ -2,21 +2,21 @@ import React, { useState, useEffect } from 'react';
 import { View, Text, TextInput, TouchableOpacity, ScrollView, Alert, StyleSheet } from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
 import { Picker } from '@react-native-picker/picker';
-import { useNavigation, useLocalSearchParams, router } from 'expo-router';
+import axios from 'axios';
+import { Tour, User } from '@/types/interface';
+import { useNavigation } from 'expo-router';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { getUsersByGuideId, getUserById, createTour } from '@/config/authApi';
-import { Tour, User } from '@/types/interface';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import CustomButton from '@/components/CustomButton';
-import { sendMessage } from '@/config/chatApi';
 
-const CreateTourScreen: React.FC = () => {
+const CreateTourScreen = () => {
   const navigation = useNavigation();
-  const { chatId, userId, nowId, userName } = useLocalSearchParams();
+
   const [user, setUser] = useState<User | null>(null);
   const [customers, setCustomers] = useState<User[]>([]);
   const [userIds, setUserIds] = useState<string[]>([]);
-  const [selectedCustomer, setSelectedCustomer] = useState<string>(userId as string);
+  const [selectedCustomer, setSelectedCustomer] = useState<string>('');
   const [form, setForm] = useState<Tour>({
     _id: '',
     user_id: '',
@@ -108,184 +108,140 @@ const CreateTourScreen: React.FC = () => {
         return { ...prevForm, Tuorlocation: newLocation };
       });
     } else if (field === 'numberUser' || field === 'price') {
-      const numericValue = parseInt(value, 10);
-      setForm(prevForm => ({
-        ...prevForm,
-        [field]: isNaN(numericValue) ? 0 : numericValue,
-      }));
+      const numericValue = parseInt(value);
+      setForm({ ...form, [field]: isNaN(numericValue) ? 0 : numericValue });
     } else {
-      setForm(prevForm => ({ ...prevForm, [field]: value }));
+      setForm({ ...form, [field]: value });
     }
   };
 
-  const showStartDatePicker = () => {
-    setStartDatePickerVisibility(true);
-  };
-
-  const showEndDatePicker = () => {
-    setEndDatePickerVisibility(true);
-  };
-
-  const hideStartDatePicker = () => {
-    setStartDatePickerVisibility(false);
-  };
-
-  const hideEndDatePicker = () => {
-    setEndDatePickerVisibility(false);
-  };
-
-  const handleStartDateChange = (event: any, selectedDate: Date | undefined) => {
-    hideStartDatePicker();
-    if (selectedDate) {
-      setForm(prevForm => ({ ...prevForm, startTime: selectedDate }));
+  const handleConfirm = async () => {
+    if (!selectedCustomer) {
+      Alert.alert('Please select a customer');
+      return;
     }
-  };
 
-  const handleEndDateChange = (event: any, selectedDate: Date | undefined) => {
-    hideEndDatePicker();
-    if (selectedDate) {
-      setForm(prevForm => ({ ...prevForm, endTime: selectedDate }));
-    }
-  };
-
-  const handleSubmit = async () => {
-    
-    const data = {
-      ...form,
-      user_id: userId as string,
-      guide_id: user?._id ?? '',
-    };
-
-    const requiredFields = ['Tuorlocation', 'schedule', 'numberUser', 'price'];
+    const requiredFields = ['Tuorlocation', 'schedule', 'numberUser', 'price', 'tourType'];
     for (const field of requiredFields) {
       const value = form[field as keyof Tour];
       if (!value || (Array.isArray(value) && value.some(item => !item))) {
-        Alert.alert('Lỗi', 'Vui lòng điền đầy đủ các trường');
+        Alert.alert('Lỗi', 'Vui lòng điền tất cả các trường');
         return;
       }
     }
 
+    const data = {
+      ...form,
+      user_id: selectedCustomer,
+      guide_id: user?._id || '',
+    };
+
     try {
-      const hello = await createTour(data);
-      const newTour : Tour = hello.tour as Tour;
+      await createTour(data);
       Alert.alert('Tour created successfully');
-
-      const tourDetails = `
-        Tour-ID: ${newTour._id}
-        Location: ${newTour.Tuorlocation.join(', ')}
-        Schedule: ${newTour.schedule}
-        Number of Users: ${newTour.numberUser}
-        Start Time: ${newTour.startTime}
-        End Time: ${newTour.endTime}
-        Price: ${newTour.price}
-        Schedule: ${newTour.schedule}
-        PAYMENT IN HERE
-      `;
-      await sendMessage(chatId as string, user?._id as string, tourDetails);
-
       navigation.goBack();
-      router.push(`/chatScreen/${chatId}?userName=${userName}&userId=${userId}`);
     } catch (error) {
       console.error('Error creating tour', error);
       Alert.alert('Error creating tour', 'An error occurred while creating the tour.');
     } finally {
-      // setForm({
-      //   _id: '',
-      //   user_id: '',
-      //   guide_id: '',
-      //   Tuorlocation: ['', ''],
-      //   schedule: '',
-      //   numberUser: 1,
-      //   startTime: new Date(Date.now()),
-      //   endTime: new Date(Date.now()),
-      //   tourType: '',
-      //   price: 100000,
-      //   status: '',
-      //   deleted: false,
-      //   createdAt: new Date(),
-      //   updatedAt: new Date(),
-      //   __v: 0,
-      // });
+      setForm({
+        _id: '',
+        user_id: '',
+        guide_id: '',
+        Tuorlocation: ['', ''],
+        schedule: '',
+        numberUser: 1,
+        startTime: new Date(Date.now()),
+        endTime: new Date(Date.now()),
+        tourType: '',
+        price: 0,
+        status: '',
+        deleted: false,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+        __v: 0,
+      });
     }
   };
 
-  React.useLayoutEffect(() => {
-    navigation.setOptions({
-      title: 'Create new Tour',
-      headerStyle: {
-        backgroundColor: '#FF8C00', // Màu nền vàng
-      },
-      headerTintColor: '#fff', // Màu chữ trắng
-      headerTitleStyle: {
-        fontWeight: 'bold' ,
-      },
-    });
-  }, [navigation]);
-
   if (isLoading || isLoading2 || isLoading3) {
     return (
-      <View className="flex-1 justify-center items-center">
-        <Text className="text-lg text-gray-500">Loading...</Text>
+      <View className='h-full w-full justify-center items-center bg-gray-100'>
+        <Text className='text-lg text-gray-600'>Loading...</Text>
       </View>
     );
   }
 
   return (
-    <SafeAreaView className="flex-1 p-4" edges={['right', 'bottom', 'left']}>
-      <ScrollView>
-        {/* <Text className="text-2xl font-semibold mb-4">Create Tour</Text> */}
-        <View className="mb-4">
-          <Text className="mb-2">Select Customer:</Text>
-          <TextInput 
-            value={userName as string}
-            editable={false}
-            className='border border-gray-300 mb-4 p-2 rounded bg-white text-black font-bold'
-          />
+    <SafeAreaView>
+      <ScrollView contentContainerStyle={styles.container}>
+        <View className=' mb-4'>
+          <Text className='text-4xl font-bold text-primary_darker text-gray-800 text-center'>Create New Tour</Text>
         </View>
 
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Location 1</Text>
-          <TextInput
-            value={form.Tuorlocation[0]}
-            onChangeText={(text) => handleInputChange('Tuorlocation', { index: 0, text })}
-            placeholder="Enter the start location"
-            className='border border-gray-300 mb-4 p-2 rounded bg-white'
-          />
-        </View>
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Customer</Text>
+        <Picker
+          selectedValue={selectedCustomer}
+          onValueChange={(itemValue) => setSelectedCustomer(itemValue)}
+          className='mb-4 bg-white'
+        >
+          <Picker.Item label="Select a customer..." value="" />
+          {customers.map(customer => (
+            <Picker.Item key={customer._id} label={customer.fullName} value={customer._id} />
+          ))}
+        </Picker>
 
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Location 2</Text>
-          <TextInput
-            value={form.Tuorlocation[1]}
-            onChangeText={(text) => handleInputChange('Tuorlocation', { index: 1, text })}
-            placeholder="Enter the destination"
-            className='border border-gray-300 mb-4 p-2 rounded bg-white'
-          />
-        </View>
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Location - FROM</Text>
+        <TextInput
+          value={form.Tuorlocation[0]}
+          onChangeText={(text) => handleInputChange('Tuorlocation', { index: 0, text })}
+          placeholder="Enter the start location"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
 
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Location - TO</Text>
+        <TextInput
+          value={form.Tuorlocation[1]}
+          onChangeText={(text) => handleInputChange('Tuorlocation', { index: 1, text })}
+          placeholder="Enter the destination"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
 
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Total Guests</Text>
-          <TextInput
-            value={form.numberUser.toString()}
-            onChangeText={(text) => handleInputChange('numberUser', text)}
-            placeholder="Enter the number of guests"
-            keyboardType="numeric"
-            className='border border-gray-300 mb-4 p-2 rounded bg-white'
-          />
-        </View>
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Total Guests</Text>
+        <TextInput
+          value={form.numberUser.toString()}
+          onChangeText={(text) => handleInputChange('numberUser', text)}
+          placeholder="Enter the number of guests"
+          keyboardType="numeric"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
 
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Start Date</Text>
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Price</Text>
+        <TextInput
+          value={form.price.toString()}
+          onChangeText={(text) => handleInputChange('price', text)}
+          placeholder="Enter the price"
+          keyboardType="numeric"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
 
-          <TouchableOpacity 
-            onPress={() => setStartDatePickerVisibility(true)} 
-            className='border border-gray-300 mb-4 p-2 rounded bg-gray-200 items-center'
-          >
-            <Text>{form.startTime.toLocaleDateString()}</Text>
-          </TouchableOpacity>
-          {isStartDatePickerVisible && (
-            <DateTimePicker
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Schedule</Text>
+        <TextInput
+          value={form.schedule}
+          onChangeText={(text) => handleInputChange('schedule', text)}
+          placeholder="Enter the schedule"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
+
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Start Date</Text>
+        <TouchableOpacity 
+          onPress={() => setStartDatePickerVisibility(true)} 
+          className='border border-gray-300 mb-4 p-2 rounded bg-gray-200 items-center'
+        >
+          <Text>{form.startTime.toLocaleDateString()}</Text>
+        </TouchableOpacity>
+        {isStartDatePickerVisible && (
+          <DateTimePicker
             value={form.startTime}
             mode="date"
             display="default"
@@ -296,23 +252,20 @@ const CreateTourScreen: React.FC = () => {
                 handleInputChange('startTime', selectedDate);
               }
             }}
-            />
-          )}
-        </View>
+          />
+        )}
 
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>End Date</Text>
-
-          <TouchableOpacity 
-            onPress={() => setEndDatePickerVisibility(true)} 
-            className='border border-gray-300 mb-4 p-2 rounded bg-gray-200 items-center'
-          >
-            <Text>
-              {form.endTime.toLocaleDateString()}
-            </Text>
-          </TouchableOpacity>
-          {isEndDatePickerVisible && (
-            <DateTimePicker
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>End Date</Text>
+        <TouchableOpacity 
+          onPress={() => setEndDatePickerVisibility(true)} 
+          className='border border-gray-300 mb-4 p-2 rounded bg-gray-200 items-center'
+        >
+          <Text>
+            {form.endTime.toLocaleDateString()}
+          </Text>
+        </TouchableOpacity>
+        {isEndDatePickerVisible && (
+          <DateTimePicker
             value={form.endTime}
             mode="date"
             display="default"
@@ -323,39 +276,22 @@ const CreateTourScreen: React.FC = () => {
                 handleInputChange('endTime', selectedDate);
               }
             }}
-            />
-          )}
-        </View>
-
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Price per person</Text>
-          <TextInput
-            value={form.price.toString()}
-            onChangeText={(text) => handleInputChange('price', text)}
-            placeholder="Enter the price"
-            keyboardType="numeric"
-            className='border border-gray-300 mb-4 p-2 rounded bg-white'
           />
-        </View>
-
-        <View className="mb-4">
-          <Text className='mb-2 text-lg font-semibold text-gray-700'>Schedule</Text>
-          <TextInput
-            value={form.schedule}
-            onChangeText={(text) => handleInputChange('schedule', text)}
-            placeholder="Enter the schedule"
-            className='border border-gray-300 mb-4 p-2 rounded bg-white h-45'
-            multiline={true}
-            numberOfLines={6}
-            textAlignVertical="top"
-          />
-        </View>
+        )}
         
+        <Text className='mb-2 text-lg font-semibold text-gray-700'>Tour Type</Text>
+        <TextInput
+          value={form.tourType}
+          onChangeText={(itemValue) => handleInputChange('tourType', itemValue)}
+          placeholder="Enter the tour type"
+          className='border border-gray-300 mb-4 p-2 rounded bg-white'
+        />
+
         <CustomButton 
-          title='Create Tour'
-          handlePress={handleSubmit} 
-          containerStyles='mt-5 py-3 rounded-full items-center bg-orange-500'
-          textStyles='text-white font-bold text-lg'
+          title='Confirm' 
+          handlePress={handleConfirm} 
+          containerStyles='p-4 rounded-lg'
+          textStyles='text-white text-center font-bold text-lg'
         />
       </ScrollView>
     </SafeAreaView>
@@ -363,18 +299,10 @@ const CreateTourScreen: React.FC = () => {
 };
 
 const styles = StyleSheet.create({
-  input: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    padding: 2,
-    borderRadius: 5,
-  },
-  picker: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-  },
+  container: {
+    padding: 16,
+    backgroundColor: '#f8f9fa',
+  }
 });
-
 
 export default CreateTourScreen;
