+++ types/point.ts
@@ -1,0 +1,6 @@
+export interface Point {
+    latitude: number;
+    longitude: number;
+    latitudeDelta: number;
+    longitudeDelta: number;
+}
